requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Faker==33.1.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.2.1
openai==1.58.1
packaging==24.2
pandas==2.2.3
phidata==2.7.5
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
pynrrd==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.4.5
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
